Move test_csv to directory
